+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * many API endpoints now require a `profile` parameter
   * initializing a profile now requires a `temp` folder parameter
   * several new endpoints and backward incompatible changes
+  * the server can store settings for the client
 
 
 ## [0.4.5] - 2024-10-17